+++ src/tr1/game/output.h
@@ -59,7 +59,7 @@ void Output_DrawSprite(
     int32_t x, int32_t y, int32_t z, int16_t sprnum, int16_t shade, RGB_F tint);
 void Output_DrawScreenSprite(
     int32_t sx, int32_t sy, int32_t z, int32_t scale_h, int32_t scale_v,
-    int32_t sprnum, int16_t shade, uint16_t flags, int32_t page);
+    int32_t sprite_idx, int16_t shade);
 void Output_DrawUISprite(
     int32_t x, int32_t y, int32_t scale, int16_t sprnum, int16_t shade);