+++ src/tr2/game/console/common.c
@@ -1,7 +1,5 @@
 #include "game/console/common.h"
 
-#include "game/text.h"
-
 void Console_DrawBackdrop(void)
 {
     // TODO: implement me