+++ src/tr2/game/output.h
@@ -50,7 +50,7 @@ void Output_DrawPickup(
 
 void Output_DrawScreenSprite(
     int32_t sx, int32_t sy, int32_t sz, int32_t scale_h, int32_t scale_v,
-    int16_t sprite_idx, int16_t shade, uint16_t flags);
+    int32_t sprite_idx, int16_t shade);
 
 void Output_ClearDepthBuffer(void);