+++ src/tr2/global/types_decomp.h
@@ -9,7 +9,6 @@
 #include <libtrx/game/items.h>
 #include <libtrx/game/math.h>
 #include <libtrx/game/rooms/types.h>
-#include <libtrx/game/text.h>
 #include <libtrx/game/types.h>
 
 #include <stdint.h>